Start splitting TracyCallstack.cpp
Main objective of starting this refactoring is to have a `SymbolResolver` abstraction with a clean interface, so that we may easily switch between implementations, and also reuse them more easily in tools and server. This should also help support non-pdb symbols on windows (MinGW mostly?) by allowing libbacktrace to be used on Windows more easily.

Move ImageCache to TracySymbols.hpp.
For now s_imageCache and s_KrnlCache are still owned by TracyCallstack.cpp but will in the end be moved per "resolver".
